+++ src/utils.ts
@@ -1,0 +1,16 @@
+// Be careful when changing typo of this list, many side effect could occurs, don't touch this list if you don't know what you are doing
+
+export const deviceTypeList: string[] = 
+[
+    'OICore',
+    'OIDiscrete',
+    'OIDiscreteVE',
+    'OIStepper',
+    'OIStepperVE',
+    'OIMixed',
+    'OIRelayLP',
+    'OIRelayHP'
+];
+
+export const sourceAddress = "http://openindus.com/oi-content/src/";
+export const binAddress = "http://openindus.com/oi-content/bin/";